backend/requirements.txt
# Backend Constructoken - Dependencies
# Python 3.9+

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client for GraphQL
httpx==0.25.1

# Environment variables
python-dotenv==1.0.0

# Date/time handling
python-dateutil==2.8.2

# JSON handling (built-in, pero incluido por claridad)
# json (built-in)

# Para testing (opcional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Para linting y formateo (opcional)
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Database (para cuando se implemente BD real)
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9  # PostgreSQL driver

# Utilities
requests==2.31.0
